Heading styling complete
